pyproject.toml
[project]
name = "avcp"
version = "2.1.0"
description = "Autonomous Venue Crowd Processing — edge ingestion layer"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-pubsub>=2.21.0",
    "numpy>=1.26.0",
    "pyzmq>=25.1.0",
    "scipy>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "C90",  # mccabe (cyclomatic complexity)
]

[tool.ruff.lint.mccabe]
max-complexity = 8

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_reexport = true

[[tool.mypy.overrides]]
module = ["scipy.*", "zmq.*"]
ignore_missing_imports = true
